优化测试流程

+++ src/test/java/org/pomo/toasterfx/test/ToasterFXTest.java
@@ -15,19 +15,12 @@
  */
 package org.pomo.toasterfx.test;
 
-import javafx.collections.ObservableList;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-import javafx.stage.Window;
 import javafx.util.Duration;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.*;
 import org.junit.runners.MethodSorters;
 import org.pomo.toasterfx.ToastBarToasterService;
-import org.pomo.toasterfx.ToasterFactory;
-import org.pomo.toasterfx.ToasterWindow;
 import org.pomo.toasterfx.model.ReferenceType;
 import org.pomo.toasterfx.model.ToastParameter;
 import org.pomo.toasterfx.model.impl.SingleToast;
@@ -79,49 +72,9 @@ public class ToasterFXTest {
     @SneakyThrows
     public static void init() {
 
-        Stage stage = FxToolkit.registerPrimaryStage();
-
-        ToasterWindow window = new ToasterWindow() {
-
-            {
-                stage.setOpacity(0);
-                FxToolkit.setupScene(() -> new Scene(new Parent() {
-                }));
-            }
-
-            @Override
-            @SneakyThrows
-            public void show() {
-                FxToolkit.showStage();
-            }
-
-            @Override
-            @SneakyThrows
-            public void close() {
-                FxToolkit.hideStage();
-            }
-
-            @Override
-            public boolean isShowing() {
-                return stage.isShowing();
-            }
-
-            @Override
-            public ObservableList<String> getStylesheets() {
-                return stage.getScene().getStylesheets();
-            }
-
-            @Override
-            public Window getWindow() {
-                return stage;
-            }
-        };
-
-        ToasterFactory toasterFactory = new ToasterFactory();
-        toasterFactory.setWindow(window);
+        FxToolkit.registerPrimaryStage();
 
         toasterService = new ToastBarToasterService();
-        toasterService.setToasterFactory(toasterFactory);
         toasterService.initialize();
     }
 
@@ -131,7 +84,8 @@ public class ToasterFXTest {
     @AfterClass
     public static void destroy() {
 
-        FXUtils.smartLater(() -> toasterService.destroy());
+        if (toasterService != null)
+            FXUtils.smartLater(() -> toasterService.destroy());
 
         log.info("ToasterFX test end.");
     }
